# Optimize cv::applyColorMap() for simple case