@chacha21 2022-03-01
```
original 4.x code
100000 x cv::applyColorMap(userColor) for size(32,32) and type CV_8UC1: 413.340877 ms
100000 x cv::applyColorMap(userColor) for size(64,64) and type CV_8UC1: 1322.934941 ms
100000 x cv::applyColorMap(userColor) for size(128,128) and type CV_8UC1: 5259.299441 ms
100000 x cv::applyColorMap(userColor) for size(256,256) and type CV_8UC1: 19481.239747 ms
100000 x cv::applyColorMap(userColor) for size(512,512) and type CV_8UC1: 53744.406390 ms
100000 x cv::applyColorMap(userColor) for size(1024,1024) and type CV_8UC1: 204328.241670 ms
100000 x cv::applyColorMap(JET) for size(32,32) and type CV_8UC1: 31242.315972 ms
100000 x cv::applyColorMap(JET) for size(64,64) and type CV_8UC1: 32304.962187 ms
100000 x cv::applyColorMap(JET) for size(128,128) and type CV_8UC1: 36312.397471 ms
100000 x cv::applyColorMap(JET) for size(256,256) and type CV_8UC1: 51336.647169 ms
100000 x cv::applyColorMap(JET) for size(512,512) and type CV_8UC1: 73202.010330 ms
100000 x cv::applyColorMap(JET) for size(1024,1024) and type CV_8UC1: 227124.843895 ms

with patch
100000 x cv::applyColorMap(userColor) for size(32,32) and type CV_8UC1: 491.504566 ms	=> +19%
100000 x cv::applyColorMap(userColor) for size(64,64) and type CV_8UC1: 748.580964 ms => -43%
100000 x cv::applyColorMap(userColor) for size(128,128) and type CV_8UC1: 1429.701348 ms	=> -73%
100000 x cv::applyColorMap(userColor) for size(256,256) and type CV_8UC1: 4100.422296 ms	=> -79%
100000 x cv::applyColorMap(userColor) for size(512,512) and type CV_8UC1: 12436.729648 ms	=> -77%
100000 x cv::applyColorMap(userColor) for size(1024,1024) and type CV_8UC1: 46005.752704 ms	=> -77%
100000 x cv::applyColorMap(JET) for size(32,32) and type CV_8UC1: 44762.778158 ms	=> +43%
100000 x cv::applyColorMap(JET) for size(64,64) and type CV_8UC1: 44510.404844 ms	=> +37%
100000 x cv::applyColorMap(JET) for size(128,128) and type CV_8UC1: 43447.038011 ms	=> +19%
100000 x cv::applyColorMap(JET) for size(256,256) and type CV_8UC1: 36111.826954 ms	=>-30%
100000 x cv::applyColorMap(JET) for size(512,512) and type CV_8UC1: 44628.305535 ms	=> -40%
100000 x cv::applyColorMap(JET) for size(1024,1024) and type CV_8UC1: 79051.147638 ms	=> -65%

```

It can be noted that small sizes seem not perform as well, especially for built-in colormaps. I am not sure how to interpret that.