@chacha21 2022-03-01
After tuning parallel work, timings are improved. But it is heavilydependant on the numbers of CPU core/threads.
```

original 4.x code
100000 x cv::applyColorMap(userColor) for size(32,32) and type CV_8UC1: 413.340877 ms
100000 x cv::applyColorMap(userColor) for size(64,64) and type CV_8UC1: 1322.934941 ms
100000 x cv::applyColorMap(userColor) for size(128,128) and type CV_8UC1: 5259.299441 ms
100000 x cv::applyColorMap(userColor) for size(256,256) and type CV_8UC1: 19481.239747 ms
100000 x cv::applyColorMap(userColor) for size(512,512) and type CV_8UC1: 53744.406390 ms
100000 x cv::applyColorMap(userColor) for size(1024,1024) and type CV_8UC1: 204328.241670 ms
100000 x cv::applyColorMap(JET) for size(32,32) and type CV_8UC1: 31242.315972 ms
100000 x cv::applyColorMap(JET) for size(64,64) and type CV_8UC1: 32304.962187 ms
100000 x cv::applyColorMap(JET) for size(128,128) and type CV_8UC1: 36312.397471 ms
100000 x cv::applyColorMap(JET) for size(256,256) and type CV_8UC1: 51336.647169 ms
100000 x cv::applyColorMap(JET) for size(512,512) and type CV_8UC1: 73202.010330 ms
100000 x cv::applyColorMap(JET) for size(1024,1024) and type CV_8UC1: 227124.843895 ms

with patch
100000 x cv::applyColorMap(userColor) for size(32,32) and type CV_8UC1: 304.287106 ms	=> -26%
100000 x cv::applyColorMap(userColor) for size(64,64) and type CV_8UC1: 1093.173997 ms	=> -17%
100000 x cv::applyColorMap(userColor) for size(128,128) and type CV_8UC1: 2211.233782 ms	=> -58%
100000 x cv::applyColorMap(userColor) for size(256,256) and type CV_8UC1: 5893.139673 ms	=> -70%
100000 x cv::applyColorMap(userColor) for size(512,512) and type CV_8UC1: 17804.391272 ms	=> -67%
100000 x cv::applyColorMap(userColor) for size(1024,1024) and type CV_8UC1: 68944.028322 ms	=> -67%
100000 x cv::applyColorMap(JET) for size(32,32) and type CV_8UC1: 31083.159630 ms	=> -0.5%
100000 x cv::applyColorMap(JET) for size(64,64) and type CV_8UC1: 31872.934770 ms	=> -1.33%
100000 x cv::applyColorMap(JET) for size(128,128) and type CV_8UC1: 33071.244961 ms	=> -9%
100000 x cv::applyColorMap(JET) for size(256,256) and type CV_8UC1: 37666.202003 ms	=> -27%
100000 x cv::applyColorMap(JET) for size(512,512) and type CV_8UC1: 51594.348558 ms	=> -30%
100000 x cv::applyColorMap(JET) for size(1024,1024) and type CV_8UC1: 104773.154826 ms	=> -54%

```