@chacha21 2022-03-01
```
original 4.x
100000 x cv::applyColorMap(userColor) for size(32,32) and type CV_8UC1: 413.340877 ms
100000 x cv::applyColorMap(userColor) for size(64,64) and type CV_8UC1: 1322.934941 ms
100000 x cv::applyColorMap(userColor) for size(128,128) and type CV_8UC1: 5259.299441 ms
100000 x cv::applyColorMap(userColor) for size(256,256) and type CV_8UC1: 19481.239747 ms
100000 x cv::applyColorMap(userColor) for size(512,512) and type CV_8UC1: 53744.406390 ms
100000 x cv::applyColorMap(userColor) for size(1024,1024) and type CV_8UC1: 204328.241670 ms
100000 x cv::applyColorMap(JET) for size(32,32) and type CV_8UC1: 31242.315972 ms
100000 x cv::applyColorMap(JET) for size(64,64) and type CV_8UC1: 32304.962187 ms
100000 x cv::applyColorMap(JET) for size(128,128) and type CV_8UC1: 36312.397471 ms
100000 x cv::applyColorMap(JET) for size(256,256) and type CV_8UC1: 51336.647169 ms
100000 x cv::applyColorMap(JET) for size(512,512) and type CV_8UC1: 73202.010330 ms
100000 x cv::applyColorMap(JET) for size(1024,1024) and type CV_8UC1: 227124.843895 ms

with patch
100000 x cv::applyColorMap(userColor) for size(32,32) and type CV_8UC1: 137.762362 ms => -67%
100000 x cv::applyColorMap(userColor) for size(64,64) and type CV_8UC1: 474.281744 ms => -64%
100000 x cv::applyColorMap(userColor) for size(128,128) and type CV_8UC1: 1154.152432 ms => -78%
100000 x cv::applyColorMap(userColor) for size(256,256) and type CV_8UC1: 3046.539612 ms => -84%
100000 x cv::applyColorMap(userColor) for size(512,512) and type CV_8UC1: 8813.734257 ms => -84%
100000 x cv::applyColorMap(userColor) for size(1024,1024) and type CV_8UC1: 32218.623244 ms => -84%
100000 x cv::applyColorMap(JET) for size(32,32) and type CV_8UC1: 30845.112720 ms => -1,3%
100000 x cv::applyColorMap(JET) for size(64,64) and type CV_8UC1: 31141.354067 ms => -3.6%
100000 x cv::applyColorMap(JET) for size(128,128) and type CV_8UC1: 31967.281620 ms => -12%
100000 x cv::applyColorMap(JET) for size(256,256) and type CV_8UC1: 35685.712188 ms => -30%
100000 x cv::applyColorMap(JET) for size(512,512) and type CV_8UC1: 40515.865259 ms => -45%
100000 x cv::applyColorMap(JET) for size(1024,1024) and type CV_8UC1: 62885.093060 ms => -72%

```